+++ kurtosis-module/impl/cl_client_network/cl_client_rest_client/api_response_objects.go
@@ -1,9 +1,0 @@
-package cl_client_rest_client
-
-type GetNodeIdentityResponse struct {
-	Data *NodeIdentity `json:"data"`
-}
-
-type NodeIdentity struct {
-	ENR string	`json:"enr"`
-}