+++ kurtosis-module/impl/module.go
@@ -2,21 +2,25 @@ package impl
 
 import (
 	"encoding/json"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/cl_client_network"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/cl_client_network/teku"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/el_client_network"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/el_client_network/geth"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/forkmon"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/cl_client_rest_client"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/lodestar"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/nimbus"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/teku"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el/geth"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el/nethermind"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator/genesis_consts"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/transaction_spammer"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"path"
 	"text/template"
 	"time"
-
-	// "path"
-	// "text/template"
 )
 
 const (
@@ -25,14 +29,17 @@ const (
 	// The number of validator keys that will be preregistered inside the CL genesis file when it's created
 	numValidatorsToPreregister = 100
 
-	// NOTE: We saw issues 1 Geth node & 3 Teku nodes was causing problems, and the Teku folks
-	//  let us know that generally each CL node should be paired with 1 EL node
-	// https://discord.com/channels/697535391594446898/697539289042649190/922266717667856424
-	numElAndClPairs = 1
+	// TODO Maaaaaaaybe can't have just a single validator???? One of the Nimbus guys said that
+	numParticipants = 1
+	// numParticipants = 2
 
 	// ----------------------------------- Genesis Config Constants -----------------------------------------
-	// We COULD drop this, but it won't represent mainnet
+	// Seems to be hardcoded
+	slotsPerEpoch = uint32(32)
+
+	// If we drop this, things start to behave strangely, with slots that are of variable time lengths
 	secondsPerSlot = uint32(12)
+
 	altairForkEpoch = uint64(1)  // Set per Parithosh's recommendation
 	mergeForkEpoch = uint64(2)   // Set per Parithosh's recommendation
 	// TODO Should be set to roughly one hour (??) so that this is reached AFTER the CL gets the merge fork version (per Parithosh)
@@ -41,6 +48,9 @@ const (
 	// This is the mnemonic that will be used to generate validator keys which will be preregistered in the CL genesis.ssz that we create
 	// This is the same mnemonic that should be used to generate the validator keys that we'll load into our CL nodes when we run them
 	preregisteredValidatorKeysMnemonic = "giant issue aisle success illegal bike spike question tent bar rely arctic volcano long crawl hungry vocal artwork sniff fantasy very lucky have athlete"
+
+	// TODO What units are these?
+	genesisDelay = 0
 	// --------------------------------- End Genesis Config Constants ----------------------------------------
 
 	// ----------------------------------- Static File Constants -----------------------------------------
@@ -65,6 +75,13 @@ const (
 	// TODO uncomment these when the module can either start a private network OR connect to an existing devnet
 	// mergeDevnet3NetworkId = "1337602"
 	// mergeDevnet3ClClientBootnodeEnr = "enr:-Iq4QKuNB_wHmWon7hv5HntHiSsyE1a6cUTK1aT7xDSU_hNTLW3R4mowUboCsqYoh1kN9v3ZoSu_WuvW9Aw0tQ0Dxv6GAXxQ7Nv5gmlkgnY0gmlwhLKAlv6Jc2VjcDI1NmsxoQK6S-Cii_KmfFdUJL2TANL3ksaKUnNXvTCv1tLwXs0QgIN1ZHCCIyk"
+
+	// In normal operation, the finalized epoch will be this many epochs behind head
+	expectedNumEpochsBehindHeadForFinalizedEpoch = uint64(3)
+	firstHeadEpochWhereFinalizedEpochIsPossible = expectedNumEpochsBehindHeadForFinalizedEpoch + 1
+	timeBetweenFinalizedEpochChecks = 5 * time.Second
+	// TODO FIGURE OUT WHY THIS HAPPENS AND GET RID OF IT
+	extraDelayBeforeSlotCountStartsIncreasing = 4 * time.Minute
 )
 /*
 var mergeDevnet3BootnodeEnodes = []string{
@@ -74,18 +91,29 @@ var mergeDevnet3BootnodeEnodes = []string{
 }
  */
 
+type ExecuteParams struct {
+	WaitForFinalization bool	`json:"waitForFinalization"`
+}
+
 type ExecuteResponse struct {
 	ForkmonPublicURL string	`json:"forkmonUrl"`
 }
 
-type ExampleExecutableKurtosisModule struct {
+type Eth2KurtosisModule struct {
 }
 
-func NewExampleExecutableKurtosisModule() *ExampleExecutableKurtosisModule {
-	return &ExampleExecutableKurtosisModule{}
+func NewEth2KurtosisModule() *Eth2KurtosisModule {
+	return &Eth2KurtosisModule{}
 }
 
-func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveContext, serializedParams string) (serializedResult string, resultError error) {
+func (e Eth2KurtosisModule) Execute(enclaveCtx *enclaves.EnclaveContext, serializedParams string) (serializedResult string, resultError error) {
+	logrus.Info("Deserializing execute params...")
+	paramsObj := new(ExecuteParams)
+	if err := json.Unmarshal([]byte(serializedParams), paramsObj); err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred deserializing the serialized params")
+	}
+	logrus.Info("Successfully deserialized execute params")
+
 	logrus.Info("Generating prelaunch data...")
 	genesisUnixTimestamp := time.Now().Unix()
 	gethGenesisConfigTemplate, err := parseTemplate(gethGenesisGenerationConfigYmlTemplateFilepath)
@@ -100,6 +128,10 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 	if err != nil {
 		return "", stacktrace.Propagate(err, "An error occurred parsing the CL mnemonics YAML template")
 	}
+	nethermindGenesisJsonTemplate, err := parseTemplate(nethermindGenesisJsonTemplateFilepath)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred parsing the Nethermind genesis json template")
+	}
 	prelaunchData, err := prelaunch_data_generator.GeneratePrelaunchData(
 		enclaveCtx,
 		gethGenesisConfigTemplate,
@@ -107,8 +139,9 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 		clGenesisMnemonicsYmlTemplate,
 		preregisteredValidatorKeysMnemonic,
 		numValidatorsToPreregister,
-		numElAndClPairs,
+		numParticipants,
 		genesisUnixTimestamp,
+		genesisDelay,
 		networkId,
 		secondsPerSlot,
 		altairForkEpoch,
@@ -121,63 +154,65 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 	}
 	logrus.Info("Successfully generated prelaunch data")
 
-	// TODO Nethermind template-filling here
-	/*
-	tmpl, err := template.New(templateFilename).ParseFiles(templateFilepath)
-	template.New(
-		// For some reason, the template name has to match the basename of the file:
-		//  https://stackoverflow.com/questions/49043292/error-template-is-an-incomplete-or-empty-template
-		path.Base(nethermindGenesisJsonTemplateFilepath),
-	).Parse(
-		gethGenesisJsonFilepath,
-	)
-	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the Nethermind genesis JSON template file '%v'", nethermindGenesisJsonTemplateFilepath)
+	logrus.Info("Creating EL & CL client launchers...")
+	elClientLaunchers := map[participant_network.ParticipantELClientType]el.ELClientLauncher{
+		participant_network.ParticipantELClientType_Geth: geth.NewGethELClientLauncher(
+			prelaunchData.GethELGenesisJsonFilepathOnModuleContainer,
+			genesis_consts.PrefundedAccounts,
+		),
+		participant_network.ParticipantELClientType_Nethermind: nethermind.NewNethermindELClientLauncher(
+			nethermindGenesisJsonTemplate,
+			totalTerminalDifficulty,
+		),
 	}
-	 */
+	clGenesisPaths := prelaunchData.CLGenesisPaths
+	clClientLaunchers := map[participant_network.ParticipantCLClientType]cl.CLClientLauncher{
+		participant_network.ParticipantCLClientType_Teku: teku.NewTekuCLClientLauncher(
+			clGenesisPaths.GetConfigYMLFilepath(),
+			clGenesisPaths.GetGenesisSSZFilepath(),
+		),
+		participant_network.ParticipantCLClientType_Nimbus: nimbus.NewNimbusLauncher(
+			clGenesisPaths.GetParentDirpath(),
+		),
+		participant_network.ParticipantCLClientType_Lodestar: lodestar.NewLodestarCLClientLauncher(
+			clGenesisPaths.GetConfigYMLFilepath(),
+			clGenesisPaths.GetGenesisSSZFilepath(),
+	    ),
+	}
+	logrus.Info("Successfully created EL & CL client launchers")
 
-	logrus.Info("Launching a network of EL clients...")
-	gethClientLauncher := geth.NewGethELClientLauncher(prelaunchData.GethELGenesisJsonFilepathOnModuleContainer)
-	elNetwork := el_client_network.NewExecutionLayerNetwork(
+	logrus.Infof("Adding %v participants...", numParticipants)
+	keystoresGenerationResult := prelaunchData.KeystoresGenerationResult
+	network := participant_network.NewParticipantNetwork(
 		enclaveCtx,
 		networkId,
-		gethClientLauncher,
-	)
-
-	allElClientContexts := []*el_client_network.ExecutionLayerClientContext{}
-	for i := 0; i < numElAndClPairs; i++ {
-		elClientCtx, err := elNetwork.AddNode()
-		if err != nil {
-			return "", stacktrace.Propagate(err, "An error occurred adding EL client node %v", i)
-		}
-		allElClientContexts = append(allElClientContexts, elClientCtx)
-	}
-	logrus.Info("Successfully launched a network of EL clients")
-
-	logrus.Info("Launching a network of CL clients...")
-	clGenesisPaths := prelaunchData.CLGenesisPaths
-	// clClientLauncher := lighthouse.NewLighthouseCLClientLauncher(clGenesisPaths.GetParentDirpath())
-	clClientLauncher := teku.NewTekuCLClientLauncher(
-		clGenesisPaths.GetConfigYMLFilepath(),
-		clGenesisPaths.GetGenesisSSZFilepath(),
-	)
-	keystoresGenerationResult := prelaunchData.KeystoresGenerationResult
-	clNetwork := cl_client_network.NewConsensusLayerNetwork(
-		enclaveCtx,
-		allElClientContexts,
-		clClientLauncher,
 		keystoresGenerationResult.PerNodeKeystoreDirpaths,
+		elClientLaunchers,
+		clClientLaunchers,
 	)
 
-	allClClientContexts := []*cl_client_network.ConsensusLayerClientContext{}
-	for i := 0; i < numElAndClPairs; i++ {
-		clClientCtx, err := clNetwork.AddNode()
+	allElClientContexts := []*el.ELClientContext{}
+	allClClientContexts := []*cl.CLClientContext{}
+	for i := 0; i < numParticipants; i++ {
+		participant, err := network.AddParticipant(
+			participant_network.ParticipantELClientType_Geth,
+			participant_network.ParticipantCLClientType_Nimbus,
+		)
 		if err != nil {
-			return "", stacktrace.Propagate(err, "An error occurred adding CL client node %v", i)
+			return "", stacktrace.Propagate(err, "An error occurred adding participant %v", i)
 		}
-		allClClientContexts = append(allClClientContexts, clClientCtx)
+		allElClientContexts = append(allElClientContexts, participant.GetELClientContext())
+		allClClientContexts = append(allClClientContexts, participant.GetCLClientContext())
 	}
-	logrus.Info("Successfully launched a network of CL clients")
+	logrus.Infof("Successfully added %v partitipcants", numParticipants)
+
+
+	logrus.Info("Launching transaction spammer...")
+	// TODO Upgrade the transaction spammer so it can take in multiple EL client addresses
+	if err := transaction_spammer.LaunchTransanctionSpammer(enclaveCtx, genesis_consts.PrefundedAccounts, allElClientContexts[0]); err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred launching the transaction spammer")
+	}
+	logrus.Info("Successfully launched transaction spammer")
 
 	logrus.Info("Launching forkmon...")
 	forkmonConfigTemplate, err := parseTemplate(forkmonConfigTemplateFilepath)
@@ -192,6 +227,16 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 		secondsPerSlot,
 	)
 	logrus.Info("Successfully launched forkmon")
+
+	if paramsObj.WaitForFinalization {
+		logrus.Info("Waiting for the first finalized epoch...")
+		firstClClientCtx := allClClientContexts[0]
+		firstClClientRestClient := firstClClientCtx.GetRESTClient()
+		if err := waitUntilFirstFinalizedEpoch(firstClClientRestClient); err != nil {
+			return "", stacktrace.Propagate(err, "An error occurred waiting until the first finalized epoch occurred")
+		}
+		logrus.Info("First finalized epoch occurred successfully")
+	}
 
 	responseObj := &ExecuteResponse{
 		ForkmonPublicURL: forkmonPublicUrl,
@@ -216,4 +261,35 @@ func parseTemplate(filepath string) (*template.Template, error) {
 		return nil, stacktrace.Propagate(err, "An error occurred parsing template file '%v'", filepath)
 	}
 	return tmpl, nil
+}
+
+func waitUntilFirstFinalizedEpoch(restClient *cl_client_rest_client.CLClientRESTClient) error {
+	// If we wait long enough that we might be in this epoch, we've waited too long - finality should already have happened
+	waitedTooLongEpoch := firstHeadEpochWhereFinalizedEpochIsPossible + 1
+	timeoutSeconds := waitedTooLongEpoch * uint64(slotsPerEpoch) * uint64(secondsPerSlot)
+	timeout := time.Duration(timeoutSeconds) * time.Second + extraDelayBeforeSlotCountStartsIncreasing
+	deadline := time.Now().Add(timeout)
+
+	for time.Now().Before(deadline) {
+		currentSlot, err := restClient.GetCurrentSlot()
+		if err != nil {
+			return stacktrace.Propagate(err, "An error occurred getting the current slot using the REST client, which should never happen")
+		}
+		currentEpoch := currentSlot / uint64(slotsPerEpoch)
+		finalizedEpoch, err := restClient.GetFinalizedEpoch()
+		if err != nil {
+			return stacktrace.Propagate(err, "An error occurred getting the finalized epoch using the REST client, which should never happen")
+		}
+		if finalizedEpoch > 0 && finalizedEpoch + expectedNumEpochsBehindHeadForFinalizedEpoch == currentEpoch {
+			return nil
+		}
+		logrus.Debugf(
+			"Finalized epoch hasn't occurred yet; current slot = '%v', current epoch = '%v', and finalized epoch = '%v'",
+			currentSlot,
+			currentEpoch,
+			finalizedEpoch,
+		 )
+		time.Sleep(timeBetweenFinalizedEpochChecks)
+	}
+	return stacktrace.NewError("Waited for %v for the finalized epoch to be %v epochs behind the current epoch, but it didn't happen", timeout, expectedNumEpochsBehindHeadForFinalizedEpoch)
 }