+++ kurtosis-module/impl/participant_network/el/el_client_context.go
@@ -1,0 +1,30 @@
+package el
+
+type ELClientContext struct {
+	enr string
+	enode string
+	ipAddr string
+	rpcPortNum  uint16
+	wsPortNum uint16
+}
+
+func NewELClientContext(enr string, enode string, ipAddr string, rpcPortNum uint16, wsPortNum uint16) *ELClientContext {
+	return &ELClientContext{enr: enr, enode: enode, ipAddr: ipAddr, rpcPortNum: rpcPortNum, wsPortNum: wsPortNum}
+}
+
+func (ctx *ELClientContext) GetENR() string {
+	return ctx.enr
+}
+func (ctx *ELClientContext) GetEnode() string {
+	return ctx.enode
+}
+func (ctx *ELClientContext) GetIPAddress() string {
+	return ctx.ipAddr
+}
+func (ctx *ELClientContext) GetRPCPortNum() uint16 {
+	return ctx.rpcPortNum
+}
+func (ctx *ELClientContext) GetWSPortNum() uint16 {
+	return ctx.wsPortNum
+}
+