+++ kurtosis-module/impl/participant_network/el/geth/geth_launcher.go
@@ -4,7 +4,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/el_client_network"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator/genesis_consts"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/service_launch_utils"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
@@ -12,11 +13,14 @@ import (
 	"github.com/sirupsen/logrus"
 	"io/ioutil"
 	"net/http"
+	"os"
+	"path"
 	"strings"
 	"time"
 )
 
 const (
+	// TODO CHANGE THIS
 	imageName = "kurtosistech/go-ethereum:d99ac5a7d"
 
 	rpcPortNum       uint16 = 8545
@@ -39,17 +43,22 @@ const (
 
 	// The dirpath of the execution data directory on the client container
 	executionDataDirpathOnClientContainer = "/execution-data"
+	keystoreDirpathOnClientContainer = executionDataDirpathOnClientContainer + "/keystore"
 
+	gethKeysRelDirpathInSharedDir = "geth-keys"
 
 	jsonContentTypeHeader = "application/json"
 	rpcRequestTimeout = 5 * time.Second
 
 	getNodeInfoRpcRequestBody = `{"jsonrpc":"2.0","method": "admin_nodeInfo","params":[],"id":1}`
-	getNodeInfoMaxRetries = 10
-	getNodeInfoTimeBetweenRetries = 500 * time.Millisecond
 
-	// To start a bootnode rather than a child node, we provide this string to the launchNode function
-	bootnodeEnodeStrForStartingBootnode = ""
+	expectedSecondsForGethInit = 5
+	expectedSecondsPerKeyImport = 8
+	expectedSecondsAfterNodeStartUntilHttpServerIsAvailable = 10
+	getNodeInfoTimeBetweenRetries = 1 * time.Second
+
+	gethAccountPassword      = "password"          // Password that the Geth accounts will be locked with
+	gethAccountPasswordsFile = "/tmp/password.txt" // Importing an account to
 )
 var usedPorts = map[string]*services.PortSpec{
 	rpcPortId:          services.NewPortSpec(rpcPortNum, services.PortProtocol_TCP),
@@ -58,83 +67,54 @@ var usedPorts = map[string]*services.PortSpec{
 	udpDiscoveryPortId: services.NewPortSpec(discoveryPortNum, services.PortProtocol_UDP),
 }
 var entrypointArgs = []string{"sh", "-c"}
+var prefundedAccountPrivateKeys = []string{
+	"ef5177cd0b6b21c87db5a0bf35d4084a8a57a9d6a064f86d51ac85f2b873a4e2",  // m/44'/60'/0'/0/0
+	"48fcc39ae27a0e8bf0274021ae6ebd8fe4a0e12623d61464c498900b28feb567", // m/44'/60'/0'/0/1
+	"7988b3a148716ff800414935b305436493e1f25237a2a03e5eebc343735e2f31", // m/44'/60'/0'/0/2
+	"b3c409b6b0b3aa5e65ab2dc1930534608239a478106acf6f3d9178e9f9b00b35", // m/44'/60'/0'/0/3
+	"df9bb6de5d3dc59595bcaa676397d837ff49441d211878c024eabda2cd067c9f", // m/44'/60'/0'/0/4
+	"7da08f856b5956d40a72968f93396f6acff17193f013e8053f6fbb6c08c194d6", // m/44'/60'/0'/0/5
+}
 
 type GethELClientLauncher struct {
 	genesisJsonFilepathOnModuleContainer string
+	prefundedAccountInfo []*genesis_consts.PrefundedAccount
 }
 
-func NewGethELClientLauncher(genesisJsonFilepathOnModuleContainer string) *GethELClientLauncher {
-	return &GethELClientLauncher{genesisJsonFilepathOnModuleContainer: genesisJsonFilepathOnModuleContainer}
+func NewGethELClientLauncher(genesisJsonFilepathOnModuleContainer string, prefundedAccountInfo []*genesis_consts.PrefundedAccount) *GethELClientLauncher {
+	return &GethELClientLauncher{genesisJsonFilepathOnModuleContainer: genesisJsonFilepathOnModuleContainer, prefundedAccountInfo: prefundedAccountInfo}
 }
 
-func (launcher *GethELClientLauncher) LaunchBootNode(
-	enclaveCtx *enclaves.EnclaveContext,
-	serviceId services.ServiceID,
-	networkId string,
-) (
-	resultClientCtx *el_client_network.ExecutionLayerClientContext,
-	resultErr error,
-) {
-	clientCtx, err := launcher.launchNode(enclaveCtx, serviceId, networkId, bootnodeEnodeStrForStartingBootnode)
+func (launcher *GethELClientLauncher) Launch(enclaveCtx *enclaves.EnclaveContext, serviceId services.ServiceID, networkId string, bootnodeContext *el.ELClientContext) (resultClientCtx *el.ELClientContext, resultErr error) {
+	containerConfigSupplier := launcher.getContainerConfigSupplier(networkId, bootnodeContext)
+	serviceCtx, err := enclaveCtx.AddService(serviceId, containerConfigSupplier)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred starting boot Geth node with service ID '%v'", serviceId)
+		return nil, stacktrace.Propagate(err, "An error occurred launching the Geth EL client with service ID '%v'", serviceId)
 	}
-	return clientCtx, nil
-}
 
-func (launcher *GethELClientLauncher) LaunchChildNode(
-	enclaveCtx *enclaves.EnclaveContext,
-	serviceId services.ServiceID,
-	networkId string,
-	bootnodeEnode string,
-) (
-	resultClientCtx *el_client_network.ExecutionLayerClientContext,
-	resultErr error,
-) {
-	clientCtx, err := launcher.launchNode(enclaveCtx, serviceId, networkId, bootnodeEnode)
+	nodeInfo, err := launcher.getNodeInfoWithRetry(serviceCtx.GetPrivateIPAddress())
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred starting child Geth node with service ID '%v' connected to boot node with enode '%v'", serviceId, bootnodeEnode)
+		return nil, stacktrace.Propagate(err, "An error occurred getting the newly-started node's info")
 	}
-	return clientCtx, nil
+
+	result := el.NewELClientContext(
+		nodeInfo.ENR,
+		nodeInfo.Enode,
+		serviceCtx.GetPrivateIPAddress(),
+		rpcPortNum,
+		wsPortNum,
+	)
+
+	return result, nil
 }
 
 
 // ====================================================================================================
 //                                       Private Helper Methods
 // ====================================================================================================
-func (launcher *GethELClientLauncher) launchNode(
-	enclaveCtx *enclaves.EnclaveContext,
-	serviceId services.ServiceID,
-	networkId string,
-	bootnodeEnode string, // NOTE: If this is emptystring, the node will be launched as a bootnode
-) (
-	resultClientCtx *el_client_network.ExecutionLayerClientContext,
-	resultErr error,
-) {
-	containerConfigSupplier := launcher.getContainerConfigSupplier(networkId, bootnodeEnode)
-	serviceCtx, err := enclaveCtx.AddService(serviceId, containerConfigSupplier)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred launching the Geth EL client with service ID '%v'", serviceId)
-	}
-
-	nodeInfo, err := getNodeInfoWithRetry(serviceCtx.GetPrivateIPAddress())
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting the newly-started node's info")
-	}
-
-	result := el_client_network.NewExecutionLayerClientContext(
-		serviceCtx,
-		nodeInfo.ENR,
-		nodeInfo.Enode,
-		rpcPortId,
-	)
-
-	return result, nil
-}
-
 func (launcher *GethELClientLauncher) getContainerConfigSupplier(
 	networkId string,
-	bootnodeEnode string, // NOTE: If this is emptystring, the node will be configured as a bootnode
+	bootnodeContext *el.ELClientContext, // NOTE: If this is empty, the node will be configured as a bootnode
 ) func(string, *services.SharedPath) (*services.ContainerConfig, error) {
 	result := func(privateIpAddr string, sharedDir *services.SharedPath) (*services.ContainerConfig, error) {
 		genesisJsonSharedPath := sharedDir.GetChildPath(sharedGenesisJsonRelFilepath)
@@ -142,13 +122,48 @@ func (launcher *GethELClientLauncher) getContainerConfigSupplier(
 			return nil, stacktrace.Propagate(err, "An error occurred copying genesis JSON file '%v' into shared directory path '%v'", launcher.genesisJsonFilepathOnModuleContainer, sharedGenesisJsonRelFilepath)
 		}
 
-		commandArgs := []string{
-			"geth",
-			"init",
-			"--datadir=" + executionDataDirpathOnClientContainer,
+		gethKeysDirSharedPath := sharedDir.GetChildPath(gethKeysRelDirpathInSharedDir)
+		if err := os.Mkdir(gethKeysDirSharedPath.GetAbsPathOnThisContainer(), os.ModePerm); err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred creating the Geth keys directory in the shared dir")
+		}
+
+		accountAddressesToUnlock := []string{}
+		for _, prefundedAccount := range launcher.prefundedAccountInfo {
+			keyFilepathOnModuleContainer := prefundedAccount.GethKeyFilepath
+			keyFilename := path.Base(keyFilepathOnModuleContainer)
+			keyRelFilepathInSharedDir := path.Join(gethKeysRelDirpathInSharedDir, keyFilename)
+			keyFileSharedPath := sharedDir.GetChildPath(keyRelFilepathInSharedDir)
+			if err := service_launch_utils.CopyFileToSharedPath(keyFilepathOnModuleContainer, keyFileSharedPath); err != nil {
+				return nil, stacktrace.Propagate(err, "An error occurred copying key file '%v' to the shared directory", keyFilepathOnModuleContainer)
+			}
+
+			accountAddressesToUnlock = append(accountAddressesToUnlock, prefundedAccount.Address)
+		}
+
+		initDatadirCmdStr := fmt.Sprintf(
+			"geth init --datadir=%v %v",
+			executionDataDirpathOnClientContainer,
 			genesisJsonSharedPath.GetAbsPathOnServiceContainer(),
-			"&&",
+		)
+
+		copyKeysIntoKeystoreCmdStr := fmt.Sprintf(
+			"cp -r %v/* %v/",
+			gethKeysDirSharedPath.GetAbsPathOnServiceContainer(),
+			keystoreDirpathOnClientContainer,
+		)
+
+		createPasswordsFileCmdStr := fmt.Sprintf(
+			"{ for i in $(seq 1 %v); do echo \"%v\" >> %v; done; }",
+			len(launcher.prefundedAccountInfo),
+			gethAccountPassword,
+			gethAccountPasswordsFile,
+		)
+
+		accountsToUnlockStr := strings.Join(accountAddressesToUnlock, ",")
+		launchNodeCmdArgs := []string{
 			"geth",
+			"--unlock=" + accountsToUnlockStr,
+			"--password=" + gethAccountPasswordsFile,
 			"--mine",
 			"--miner.etherbase=" + miningRewardsAccount,
 			fmt.Sprintf("--miner.threads=%v", numMiningThreads),
@@ -161,18 +176,28 @@ func (launcher *GethELClientLauncher) getContainerConfigSupplier(
 			//  that users should NOT store private information in these Kurtosis nodes!
 			"--http.api=admin,engine,net,eth",
 			"--ws",
+			"--ws.addr=0.0.0.0",
+			fmt.Sprintf("--ws.port=%v", wsPortNum),
 			"--ws.api=engine,net,eth",
 			"--allow-insecure-unlock",
 			"--nat=extip:" + privateIpAddr,
 			"--verbosity=3",
 		}
-		if bootnodeEnode != bootnodeEnodeStrForStartingBootnode {
-			commandArgs = append(
-				commandArgs,
-				"--bootnodes=" + bootnodeEnode,
+		if bootnodeContext != nil {
+			launchNodeCmdArgs = append(
+				launchNodeCmdArgs,
+				"--bootnodes=" + bootnodeContext.GetEnode(),
 			)
 		}
-		commandStr := strings.Join(commandArgs, " ")
+		launchNodeCmdStr := strings.Join(launchNodeCmdArgs, " ")
+
+		subcommandStrs := []string{
+			initDatadirCmdStr,
+			copyKeysIntoKeystoreCmdStr,
+			createPasswordsFileCmdStr,
+			launchNodeCmdStr,
+		}
+		commandStr := strings.Join(subcommandStrs, " && ")
 
 		containerConfig := services.NewContainerConfigBuilder(
 			imageName,
@@ -189,9 +214,11 @@ func (launcher *GethELClientLauncher) getContainerConfigSupplier(
 	return result
 }
 
-func getNodeInfoWithRetry(privateIpAddr string) (NodeInfo, error) {
+func (launcher *GethELClientLauncher) getNodeInfoWithRetry(privateIpAddr string) (NodeInfo, error) {
+	maxNumRetries := expectedSecondsForGethInit + len(launcher.prefundedAccountInfo) * expectedSecondsPerKeyImport + expectedSecondsAfterNodeStartUntilHttpServerIsAvailable
+
 	getNodeInfoResponse := new(GetNodeInfoResponse)
-	for i := 0; i < getNodeInfoMaxRetries; i++ {
+	for i := 0; i < maxNumRetries; i++ {
 		if err := sendRpcCall(privateIpAddr, getNodeInfoRpcRequestBody, getNodeInfoResponse); err == nil {
 			return getNodeInfoResponse.Result, nil
 		} else {
@@ -199,7 +226,7 @@ func getNodeInfoWithRetry(privateIpAddr string) (NodeInfo, error) {
 		}
 		time.Sleep(getNodeInfoTimeBetweenRetries)
 	}
-	return NodeInfo{}, stacktrace.NewError("Couldn't get the node's info even after %v retries with %v between retries", getNodeInfoMaxRetries, getNodeInfoTimeBetweenRetries)
+	return NodeInfo{}, stacktrace.NewError("Couldn't get the node's info even after %v retries with %v between retries", maxNumRetries, getNodeInfoTimeBetweenRetries)
 }
 
 func sendRpcCall(privateIpAddr string, requestBody string, targetStruct interface{}) error {