+++ kurtosis-module/impl/participant_network/el/random_transaction_sender.go
@@ -1,0 +1,5 @@
+package el
+
+type RandomTransactionSender interface {
+	SendRandomTransaction() error
+}