+++ kurtosis-module/main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
-	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/lib/execution"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl"
+	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/lib/execution"
 	"github.com/sirupsen/logrus"
 	"os"
 )
@@ -14,12 +14,7 @@ const (
 )
 
 func main() {
-
-	// >>>>>>>>>>>>>>>>>>> REPLACE WITH YOUR OWN CONFIGURATOR <<<<<<<<<<<<<<<<<<<<<<<<
-	configurator := impl.NewExampleExecutableKurtosisModuleConfigurator()
-	// >>>>>>>>>>>>>>>>>>> REPLACE WITH YOUR OWN CONFIGURATOR <<<<<<<<<<<<<<<<<<<<<<<<
-
-
+	configurator := impl.NewEth2KurtosisModuleConfigurator()
 	executor := execution.NewKurtosisModuleExecutor(configurator)
 	if err := executor.Run(); err != nil {
 		logrus.Errorf("An error occurred running the Kurtosis module executor:")