+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # TBD
 
+# 0.2.2
+### Features
+* Added a functioning Nimbus CL client
+* Added Nethermind EL client
+* Added a Lodestar CL beacon Launcher
+* Added a transaction spammer to blast the network with transactions after all the nodes come up
+* Added optional waiting until epoch finalization occurs
+
 # 0.2.1
 * Empty commit to try and kick CircleCI into actually building the tag
 
@@ -17,6 +25,10 @@
 
 ### Fixes
 * Correct merge parameters like TTD, Altair fork version, merge fork version, etc. per Parithosh's recommendations
+* Give Teku nodes 120s to start
+
+### Changes
+* Refactor the structure to reflect that there should be one EL node per CL node (and prepare for separated Beacon/validator nodes, like Lighthouse does)
 
 # 0.1.0
 * Initial commit